cloud_functions/scene_parser/requirements.txt
functions-framework==3.0.0
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.38.0
